move embedding folder
