+++ web/src/app/admin/models/embedding/CloudEmbeddingPage.tsx
