+++ web/src/app/admin/models/embedding/OpenEmbeddingPage.tsx
