+++ web/src/app/admin/models/embedding/components/ModelSelector.tsx
