+++ web/src/app/admin/models/embedding/components/ReindexingProgressTable.tsx
