+++ web/src/app/admin/models/embedding/components/types.ts
