+++ web/src/app/admin/models/embedding/modals/AlreadyPickedModal.tsx
