+++ web/src/app/admin/models/embedding/modals/ChangeCredentialsModal.tsx
