+++ web/src/app/admin/models/embedding/modals/DeleteCredentialsModal.tsx
