+++ web/src/app/admin/models/embedding/modals/ModelNotConfiguredModal.tsx
