+++ web/src/app/admin/models/embedding/modals/ModelSelectionModal.tsx
