+++ web/src/app/admin/models/embedding/modals/ProviderCreationModal.tsx
