+++ web/src/app/admin/models/embedding/modals/SelectModelModal.tsx
