+++ web/src/app/admin/models/llm/constants.ts
@@ -1,1 +1,4 @@
 export const LLM_PROVIDERS_ADMIN_URL = "/api/admin/llm/provider";
+
+export const EMBEDDING_PROVIDERS_ADMIN_URL =
+  "/api/admin/embedding/embedding-provider";
