--- web/src/app/admin/models/models/llm/constants.ts
@@ -1,3 +1,0 @@
-export const LLM_PROVIDERS_ADMIN_URL = "/api/admin/llm/provider";
-export const EMBEDDING_PROVIDERS_ADMIN_URL =
-  "/api/admin/embedding/embedding-provider";
